+++ tests/test_bam.py
@@ -4,17 +4,99 @@ import os
 import shutil
 import pytest
 import sctoolbox.bam
+import sctoolbox.tools.bam as stb
 import glob
 import scanpy as sc
+import logging
+import random
+import re
 
 
 @pytest.fixture
-def bam_handle():
+def bam_file():
+    """Fixture pointing to test bam."""
+    return os.path.join(os.path.dirname(__file__), 'data', 'atac', 'mm10_atac.bam')
+
+
+@pytest.fixture
+def bam_handle(bam_file):
     """Fixture for a bam file handle."""
-    bam_f = os.path.join(os.path.dirname(__file__), 'data', 'atac', 'mm10_atac.bam')
-    handle = sctoolbox.bam.open_bam(bam_f, "rb")
+    handle = sctoolbox.bam.open_bam(bam_file, "rb")
 
     return handle
+
+
+@pytest.fixture
+def barcodes(bam_file, bam_handle):
+    """Return 100 randomly selected barcodes."""
+    # bam_handle is not used to avoid consuming the iterator
+    read_count = stb.open_bam(bam_file, "rb").count()
+
+    # select random indexes
+    indices = random.sample(range(0, read_count), k=100)
+
+    return [read.get_tag("CB") for index, read in enumerate(bam_handle) if index in indices]
+
+
+@pytest.fixture(scope="session")
+def adata():
+    """Load and returns an anndata object."""
+
+    # has .X of type numpy.array
+    obj = sc.read_h5ad(os.path.join(os.path.dirname(__file__), 'data', 'atac', 'mm10_atac.h5ad'))
+
+    return obj
+
+
+def test_check_barcode_tag(adata, bam_file, mocker, caplog):
+    """Tests the barcode overlap amount between adata and bam file."""
+    caplog.set_level(logging.INFO)
+
+    # test overlap == 0%
+    mocker.patch('sctoolbox.tools.bam.bam_adata_ov', return_value=0)
+    stb.check_barcode_tag(adata=adata, bamfile=bam_file, cb_tag="CB")
+    assert 'None of the barcodes from the bamfile found in the .obs table.\nConsider if you are using the wrong column cb-tag or bamfile.' in caplog.text
+
+    # test overlap <= 5%
+    mocker.patch('sctoolbox.tools.bam.bam_adata_ov', return_value=0.05)
+    stb.check_barcode_tag(adata=adata, bamfile=bam_file, cb_tag="CB")
+    assert 'Only 5% or less of the barcodes from the bamfile found in the .obs table.\nConsider if you are using the wrong column for cb-tag or bamfile.' in caplog.text
+
+    # test overlap > 5%
+    mocker.patch('sctoolbox.tools.bam.bam_adata_ov', return_value=0.8)
+    stb.check_barcode_tag(adata=adata, bamfile=bam_file, cb_tag="CB")
+    assert 'Barcode tag: OK' in caplog.text
+
+    # test overlap error (TODO don't know how this could be triggered)
+    with pytest.raises(ValueError):
+        mocker.patch('sctoolbox.tools.bam.bam_adata_ov', return_value=float("nan"))
+        stb.check_barcode_tag(adata=adata, bamfile=bam_file, cb_tag="CB")
+
+
+def test_subset_bam(bam_file, barcodes, caplog, tmpdir):
+    """Check the subset_bam function."""
+    outfile = tmpdir / "subset.bam"
+
+    # check success
+    stb.subset_bam(bam_in=bam_file,
+                   bam_out=str(outfile),
+                   barcodes=barcodes,
+                   read_tag="CB",
+                   pysam_threads=4,
+                   overwrite=False)
+
+    assert bool(re.match(r"Wrote \d+ reads to output bam", caplog.messages[-1]))
+    assert outfile.isfile()
+
+    # check overwrite warning
+    stb.subset_bam(bam_in=bam_file,
+                   bam_out=str(outfile),
+                   barcodes=barcodes,
+                   read_tag="CB",
+                   pysam_threads=4,
+                   overwrite=False)
+
+    assert f"Output file {str(outfile)} exists. Skipping." in caplog.text
 
 
 def test_open_bam(bam_handle):  # this is indirectly a test of sctoolbox.bam.open_bam
@@ -31,18 +113,14 @@ def test_get_bam_reads(bam_handle):
     assert total == 10000
 
 
-def test_split_bam_clusters(bam_handle):
+@pytest.mark.parametrize("parallel,sort_bams,index_bams", [(True, True, True), (False, False, False)])
+def test_split_bam_clusters(bam_handle, bam_file, adata, parallel, sort_bams, index_bams):
     """Test split_bam_clusters success."""
-
-    bam_f = os.path.join(os.path.dirname(__file__), 'data', 'atac', 'mm10_atac.bam')
-    adata_f = os.path.join(os.path.dirname(__file__), 'data', 'atac', 'mm10_atac.h5ad')
-
     # Get input reads
     n_reads_input = sctoolbox.bam.get_bam_reads(bam_handle)
 
     # Split bam
-    adata = sc.read_h5ad(adata_f)
-    sctoolbox.bam.split_bam_clusters(adata, bam_f, groupby="Sample", parallel=True)
+    sctoolbox.bam.split_bam_clusters(adata, bam_file, groupby="Sample", parallel=parallel, sort_bams=sort_bams, index_bams=index_bams, writer_threads=len(set(adata.obs["Sample"])) + 1)
 
     # Check if the bam file is split and the right size
     output_bams = glob.glob("split_Sample*.bam")
@@ -54,6 +132,21 @@ def test_split_bam_clusters(bam_handle):
     # Clean up
     for bam in output_bams:
         os.remove(bam)
+
+
+def test_failure_split_bam_clusters(bam_file, adata):
+    """Test split_bam_clusters failure."""
+    # test groupby
+    with pytest.raises(ValueError):
+        sctoolbox.bam.split_bam_clusters(adata, bam_file, groupby="SOME_NONEXISTENT_COLUMN_NAME")
+
+    # test barcode_col
+    with pytest.raises(ValueError):
+        sctoolbox.bam.split_bam_clusters(adata, bam_file, groupby="Sample", barcode_col="SOME_NONEXISTENT_TAG")
+
+    # test sort_bam/index_bam
+    with pytest.raises(ValueError):
+        sctoolbox.bam.split_bam_clusters(adata, bam_file, groupby="Sample", sort_bams=False, index_bams=True)
 
 
 def test_bam_to_bigwig():