+++ CHANGES.rst
@@ -10,6 +10,7 @@
 - added tests for _add_path() and _overlap_two_bedfiles() (#177)
 - constraint ipywidgets version to 7.7.5 to fix the quality_violinplot() (#151)(#143)
 - Add temp_dir to calc_overlap_fc.py (#167) and revised related functions
+- more testing (mainly sctoolbox.tools) (#166)
 
 
 0.3 (30-11-2023)